ASR_Models/asr_requirements.txt
absl-py==2.3.1
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.0
attrdict==2.0.1
attrs==25.4.0
audioread==3.1.0
bitsandbytes==0.46.0
braceexpand==0.1.7
cdifflib==1.2.6
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
ctc_segmentation==1.7.4
cuda-bindings==13.0.3
cuda-core==0.4.1
cuda-pathfinder==1.3.2
cuda-python==13.0.3
cycler==0.12.1
Cython==3.2.1
cytoolz==1.1.0
datasets==4.4.1
debugpy==1.8.17
decorator==5.2.1
dill==0.4.0
diskcache==5.6.3
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
executing==2.2.1
fiddle==0.3.0
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.12.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.11
indic_numtowords==1.1.0
inflect==7.5.0
intervaltree==3.1.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
Janome==0.5.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiwer==3.1.0
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kaldi-python-io==1.2.2
kaldialign==0.9.1
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
lazy_loader==0.4
lhotse==1.31.1
libcst==1.8.6
librosa==0.11.0
lightning==2.4.0
lightning-utilities==0.15.2
lilcom==1.8.1
llvmlite==0.45.1
lxml==6.0.2
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
marshmallow==4.1.0
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapy==1.1.6
ml_dtypes==0.5.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
nemo-toolkit @ git+https://github.com/NVIDIA/NeMo.git@559cc979d09f12734bc4868bf13602ff51e1ebd5
nemo_text_processing==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
num2words==0.5.14
numba==0.62.1
numba-cuda==0.20.1
numexpr==2.13.1
numpy==1.26.4
nv-one-logger-core==2.3.1
nv-one-logger-pytorch-lightning-integration==2.3.1
nv-one-logger-training-telemetry==2.3.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cccl==13.0.85
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
nvidia-nvvm==13.0.88
omegaconf==2.3.0
onnx==1.19.1
openai-whisper==20250625
optuna==4.6.0
overrides==7.7.0
packaging==24.2
pandas==2.3.3
parso==0.8.5
peft==0.18.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
portalocker==3.2.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==5.29.5
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyloudnorm==0.1.1
pynini==2.1.6.post1
pyparsing==3.2.5
pypinyin==0.55.0
pypinyin-dict==0.9.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.6
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
resampy==0.4.3
rich==14.2.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
sacrebleu==2.5.1
sacremoses==0.1.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
sentencepiece==0.2.1
sentry-sdk==2.44.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==1.0.0
SQLAlchemy==2.0.44
stack-data==0.6.3
StrEnum==0.4.15
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.21.4
toml==0.10.2
toolz==1.1.0
torch==2.8.0+cu128
torchaudio==2.8.0+cu128
torchmetrics==1.8.2
torchvision==0.23.0+cu128
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.3
triton==3.4.0
typeguard==4.4.4
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.23.0
wcwidth==0.2.14
webdataset==1.0.2
Werkzeug==3.1.3
wget==3.2
wheel==0.45.1
whisper_normalizer==0.1.12
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
